+++ server/src/main/java/wiki/ednotes/server/article/ArticleRepository.java
@@ -3,7 +3,7 @@ package wiki.ednotes.server.article;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ArticleRepository extends JpaRepository<Article, Long> {
+public interface ArticleRepository extends JpaRepository<Article, Integer> {
    List<Article> findByIsPublishedOrderByOrderInCategoryAsc(Boolean isPublished);
-   List<Article> findByCategoryIdOrderByOrderInCategoryAsc(Long categoryId);
+   List<Article> findByCategoryIdOrderByOrderInCategoryAsc(Integer categoryId);
 }
